+++ packages/cli/src/bin/app-upload.ts
@@ -2,6 +2,7 @@ import chalk from 'chalk';
 import AdmZip from 'adm-zip';
 
 import type { FusionFramework } from './framework.node.js';
+import type { FetchRequest } from '@equinor/fusion-framework-module-http/client';
 
 import { loadMetadata } from './helpers/load-bundle-metadata.js';
 
@@ -70,7 +71,7 @@ export const uploadApplication = async (
     // Create a client for the 'apps' service
     const appClient = await framework.serviceDiscovery.createClient('apps');
     // Subscribe to outgoing requests for logging and debugging
-    appClient.request$.subscribe((request) => {
+    appClient.request$.subscribe((request: FetchRequest) => {
       log?.info('🌎', 'Executing request to:', request.uri);
       log?.debug('Request:', request);
     });